day and month reports extend an abstract report
